requirements.txt
boto3~=1.28.57
flask~=2.3.2
python-dotenv~=1.0.0
botocore~=1.31.57
click~=8.1.6
PyJWT~=2.8.0
gunicorn==20.0.4
